refactor(rs-platform-wallet): move test-only batch-fresh accessor to e2e harness (PR #3609 C1)
Per lklimek's review on PR #3609: `next_unused_receive_addresses` and
its `derive_fresh_unused_addresses` helper are only used by PA-005b
(plus their own unit tests) — production flows always take one address
at a time through `next_unused_receive_address`. Mirror the precedent
in commit 468e77472c ("revert(rs-platform-wallet): drop test-only
production additions; absorb in e2e framework") and relocate the
batch-fresh path to a new `tests/e2e/framework/gap_limit.rs` module:

- `next_unused_receive_addresses(wallet, account_key, count)` — free
  function exposing the same lock-and-lookup wrapper, driven via the
  public `PlatformWallet::wallet_manager()` / `wallet_id()` accessors.
- `derive_fresh_unused_addresses(pool, key_source, count)` — pure
  pool-level helper, `pub(super)` for the unit tests that pin the
  gap-limit ceiling math (5 tests, identical assertions to the
  production unit tests they replace).

Production-side `PlatformAddressWallet::next_unused_receive_addresses`
and the supporting helper / test module are deleted. The single-address
production accessor `next_unused_receive_address` is unchanged.

Also bundles the trailing-blank-line `cargo fmt` cleanup in
framework/config.rs and an `#![allow(clippy::result_large_err)]` lint
on the e2e crate root (mirrors lib.rs's existing allow).
